Project Completed,
NEXT: Bug fixes

+++ Homework5_3/Interface/IAccount.cs
@@ -9,11 +9,11 @@ namespace Homework5_3.Interface
     internal interface IAccount
     {
         int AccountId { get; set; }
-        double Balance { get; set; }
+        decimal Balance { get; set; }
         Enums.AccountType AccountType { get; set; }
         Enums.CurrencyType CurrencyType { get; set; }
 
-        void Deposit(double amount);
-        void Withdraw(double amount);
+        void Deposit(decimal amount);
+        void Withdraw(decimal amount);
     }
 }
